AutoTranspose
Added AutoTranspose, a feature that automatically adjusts transposition before playing a song. Instead of manually spamming the up/down arrows every time, the program now sets the suggested transposition for you.

This makes life easier for degenerates (like me) who just want to sit back and let the piano play without the extra hassle.

Since transposition is ultimately personal preference, this setting is disabled by default—but for those who want a seamless experience, just turn it on and enjoy!

NOTE: this was only tested on ROBLOX pianos. If the keybind is incorrect or it doesn't work, I did allow for the transposition keys to be changed in the settings.